Create an icon

Creates an icon using based on the provided name. The icon will be returned either as a "magick-image" object (if output = "img") or as a text string containing an html img tag (if output = "tag").

The height and the width can be specified, the function will try to match the icon with the closest possible size and can resize if necessary.

icon(
  name,
  height = NULL,
  width = NULL,
  resize = FALSE,
  output = c("img", "tag")
)

Arguments

name

Icon name.

height

Icon height in pixels.

width

Icon width in pixels.

resize

Should the icon be resized to the provided height and/or width.

output

Return either a "magick-image" object or img tag text string.

Value

The return type is determined by the value of output selected.

Examples

icon("rstudio", 64)

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG       64     64 sRGB       TRUE      2422 72x72  

icon("rstudio", 72)

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG       64     64 sRGB       TRUE      2422 72x72  

icon("rstudio", 72, resize = TRUE)

#> # A tibble: 1 x 7
#>   format width height colorspace matte filesize density
#>   <chr>  <int>  <int> <chr>      <lgl>    <int> <chr>  
#> 1 PNG       72     72 sRGB       TRUE         0 72x72  

icon("rstudio", 64, output = "tag")

#> <img src="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAEAAAABACAYAAACqaXHeAAAAGXRFWHRTb2Z0d2FyZQBBZG9iZSBJbWFnZVJlYWR5ccllPAAAA3hpVFh0WE1MOmNvbS5hZG9iZS54bXAAAAAAADw/eHBhY2tldCBiZWdpbj0i77u/IiBpZD0iVzVNME1wQ2VoaUh6cmVTek5UY3prYzlkIj8+IDx4OnhtcG1ldGEgeG1sbnM6eD0iYWRvYmU6bnM6bWV0YS8iIHg6eG1wdGs9IkFkb2JlIFhNUCBDb3JlIDUuNi1jMTM4IDc5LjE1OTgyNCwgMjAxNi8wOS8xNC0wMTowOTowMSAgICAgICAgIj4gPHJkZjpSREYgeG1sbnM6cmRmPSJodHRwOi8vd3d3LnczLm9yZy8xOTk5LzAyLzIyLXJkZi1zeW50YXgtbnMjIj4gPHJkZjpEZXNjcmlwdGlvbiByZGY6YWJvdXQ9IiIgeG1sbnM6eG1wTU09Imh0dHA6Ly9ucy5hZG9iZS5jb20veGFwLzEuMC9tbS8iIHhtbG5zOnN0UmVmPSJodHRwOi8vbnMuYWRvYmUuY29tL3hhcC8xLjAvc1R5cGUvUmVzb3VyY2VSZWYjIiB4bWxuczp4bXA9Imh0dHA6Ly9ucy5hZG9iZS5jb20veGFwLzEuMC8iIHhtcE1NOk9yaWdpbmFsRG9jdW1lbnRJRD0ieG1wLmRpZDo3NmJiZDRkNi0yMzA3LTRlZDktODg3Yy0wYmE5ZWMzZjViOGEiIHhtcE1NOkRvY3VtZW50SUQ9InhtcC5kaWQ6NjdDQzY3QjhEODFDMTFFN0FBMEZFODA2MTQ0ODJDRTEiIHhtcE1NOkluc3RhbmNlSUQ9InhtcC5paWQ6NjdDQzY3QjdEODFDMTFFN0FBMEZFODA2MTQ0ODJDRTEiIHhtcDpDcmVhdG9yVG9vbD0iQWRvYmUgUGhvdG9zaG9wIENDIDIwMTcgKE1hY2ludG9zaCkiPiA8eG1wTU06RGVyaXZlZEZyb20gc3RSZWY6aW5zdGFuY2VJRD0ieG1wLmlpZDozNzhiM2FlZi1lMWU2LTQxYTYtOWI0MC1iNjQ5NDk3ZmIwYjciIHN0UmVmOmRvY3VtZW50SUQ9InhtcC5kaWQ6NzZiYmQ0ZDYtMjMwNy00ZWQ5LTg4N2MtMGJhOWVjM2Y1YjhhIi8+IDwvcmRmOkRlc2NyaXB0aW9uPiA8L3JkZjpSREY+IDwveDp4bXBtZXRhPiA8P3hwYWNrZXQgZW5kPSJyIj8+PPpmiwAABZRJREFUeNrkW1tsVFUU3VPbafqa1tIWsVIaS7HF1ActRBTwR03DhxojflQTRYP4hQ+g+Af4I8rwoX4RH2jUJoJKjAkSQvigYNWARBqcaqspbUihLXUehabUUPea2XdmOkynM9x77pwZV7KaaWd6z95rzj5n733PdWzd30uKUc1cyWxkNjCXMG9jljJL5DMBpo95kfkns4fpYXYxL6g0LlfRdZuZbcy14vRcKBHewWyJeQ9iHGJ2ME/rLAAc2Mhcz1xq4XUbhG8wf2fuY+6VWWMaORZco4y5ndnP3G2x87FYKmP0y5hl6RTAwXyBiUVkB7Oc7EO5jNkrNjjsFmAxs1OmYwWlDxViA2yps0uAdbIYPUT6ALb8KrYpEwDTzM3cz3SRfnCJbe5UQiJZAZzML5ibSX9sFludVgmAC30n+3qmoE1sdpoVAFPpE2YrZR5axXaHGQHeZj5LmQvYvutmBXiauY0yH+2JdofZBLiT+RFlDz4Un5ISADHzmVRr2YJS8cmRjADPM1dR9mGVpM0Jq0EUF++aGaW2PJ+W17poyDdJl8enZrxX6LyFbi8L7Uy1lQUz3rvovRb8nzOD43Tl2nVVIrzDPMj0zibAq8xKMyMU5OXQ8rrUo2fhvJAgjy+rou6BAH17ZkSFEJXi4854IVDM3GR2hImp6zTM36QZNNWU0Jtra4OzSQE2UaQTNWMGvGJFSds/NknuI4PB10XOHHrq/sqgQwC+2c5eb/Az8UKnqbqIVtSVUT7PIvClNdW061C/1TOhXBo37ugZgNXxZaulhuHeq5F1AK/jOW8I9333GH18/AJNToUchggQUAE2GjuCIQCalvU6LNUQoqNrKPz7kgVFKoZZLD6HBVhHGsFzaYL8V/8Nz4Kq4jwVwzwTLYB2xY5vIhI684qUNK9bDQHQim6g/x/uwu4LAR7Q0boqV/6MkFCElRDgPt2cf6QhtBUCvUNXVA51D0Zp1Cphr3PRY02RRvOJXq/K4RqxuizSwfHmhcX0YH1pOCU2EieF0x9YBAEW2OXkvTWuG4qg0oI8chXeuMrD+c9/vqTapIrc6LxYNeBoPGejMXh5go6eG1P9zYfrH1sFQHITvb8bq72x4B3pHqWjPV47I68k187RfhvwB/P92BXfWPQebiinrr/9KvsBFK8jFKA0At84pr2R9ratmG/n8IG0CwB89ctw+HU9Fz/YEWzCOAQYSrcAw+NT1NkTCY0nm6uCvQQbMIpRzuuQB2BtMDpJNobCeQjg0SUL/PqU7aHggQBndREAzRCbQ+Esrt6lUy1gcyj8BAHQwfxDJxFsCgWcRxww5tdhnQSwKRR+wA+HnBTFGZsTltSX8wto/ZrqpD/ffqAv7t/h8OuP1sStHZA4fXDM9AFS3Co7acj6I7NPp1mAdPibU8qqwT7xOXxjZJpCpy93m70y7gwZqa3pPYorQoRCvPuIJrFXfA6HAElViKToVspu/CNNoIBRDIULA+Z7lP14P7r+iV1aIcBIFjs/IgLQbAKgG7EtiwWAb2OJBAA+xfaQhc6fFN9oLgGwOuKYjC+LnPeJT9PJCAD8xdyQRQJsEJ8oWQGAA1bkBRrALb5QqgIYi8aXGew8bG9P9IG5BEDMvKhbsZQkDovt02YEAJB3PkGhp7YyBR1i85w5c7I1Ji70HHNPBji/R2xNqmBIpcjGVNpCoaMlfg0d94ttW+aa9jcrQPTu0KxZsgRbliVa7a0UwKinV8siM5pGx0fFhtWz7fOqBDBCAo+s4XjdztgcWzEw1lsy9r5UpryVAkQXUDuYtcytpPY+g0fGwFjbKerQczoFiO4nIOvC460tUlpb0Wbrk2u1yLXdZOH9TFW3x08LX6PI4/N3U+homvH4PDpPhUwcGBinUKcGRLsabfpzZMPj8/8JMACIpoexWxzsaQAAAABJRU5ErkJggg==" alt="RStudio Icon rstudio"/>